requirements.txt
streamlit==1.32.0
pinecone==6.0.2
python-dotenv==1.0.1
openai==1.12.0
pandas==2.2.1 